simplified image container for form. #2
