+++ server/public/styles.css
@@ -8,7 +8,7 @@ label {
 
 img.rounded {
   width: 100%;
-  max-height: 100%;
+  height: 100%;
   object-fit: contain;
   background-color: #e5e5e5;
 }
